requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
flask_ckeditor==0.4.6
flask_login==0.5.0
flask_mail==0.9.1
Flask_SQLAlchemy==2.5.1
flask_wtf==0.15.1
itsdangerous==2.0.1
Pillow==8.4.0
SQLAlchemy==1.4.25
WTForms==2.3.3
